more support for handling doubles.
if they try to play an unfulfilled double that they can fulfill, it won't let them. updated the random player agent so it doesn't play an unfulfilled double that it could fulfill. also added auto-play capability such that if any player (regardless of their bot's strategy) plays a double that they cannot fulfill and they draw a tile that is capable of fulfilling the double then they automatically play it.